requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
openai>=1.40.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.3.0
python-docx>=1.1.0
numpy>=1.26.0
requests>=2.31.0